third_party/blink/web_tests/platform/win/external/wpt/css/css-text/i18n/other-lang/css-text-line-break-de-in-loose-expected.txt

This is a testharness.js-based test.
[FAIL] 2024  ONE DOT LEADER may appear at line start if de and loose
  assert_approx_equals: expected 48 +/- 1 but got 78
[FAIL] 2025  TWO DOT LEADER may appear at line start if de and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 2026  HORIZONTAL ELLIPSIS may appear at line start if de and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] 22EF  MIDLINE HORIZONTAL ELLIPSIS may appear at line start if de and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
[FAIL] FE19  PRESENTATION FORM FOR VERTICAL HORIZONTAL ELLIPSIS may appear at line start if de and loose
  assert_approx_equals: expected 69 +/- 1 but got 99
Harness: the test ran to completion.
